requirements.txt
aiohttp
pyyaml
psycopg2-binary
requests